requirements.txt
twisted==15.4.0
pyopenssl==0.15.1
txJSON-RPC==0.3.1
redis==2.10.5
decorator==4.0.4
pymongo==3.1.1
zeroc-ice==3.6.1
passlib==1.6.5
requests
jpush==3.1.2
Ghost.py==0.2.3
fabric
qiniu
ndg-httpsclient
pyasn1
pingpp
BeautifulSoup
bs4
pillow
qrcode
mock==1.3.0
elasticsearch
lxml
pika
raven
flask

